Fix code to not show keys for AWS

+++ server/src/optima/parallel.py
@@ -15,18 +15,17 @@ import boto.sqs
 from boto.sqs.message import Message
 import string
 import random
+import os
 
 # route prefix: /api/parallel
 parallel = Blueprint('parallel',  __name__, static_folder = '../static')
 
 # Connect to AWS region.
-conn = boto.sqs.connect_to_region(
-    "us-west-2",
-    aws_access_key_id='AKIAI6QRN57UJ4UHE3ZQ',
-    aws_secret_access_key='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')
+conn = boto.sqs.connect_to_region( os.environ['AWS_DEFAULT_REGION'] )
     
-# Get the optima queue
-optima_queue = conn.get_queue('optima')
+# Get the optima queue. We will error out if no queue is found. This is what is desired.
+optima_queue = conn.get_all_queues('awseb-')
+optima_queue = optima_queue[0]
 
 @parallel.route('/execute/<nprocesses>/<someparameter>', methods=['GET'])
 def parallel_execution( nprocesses, someparameter ):
@@ -39,7 +38,6 @@ def parallel_execution( nprocesses, someparameter ):
     response_queue_id = id_generator()
     q=conn.create_queue(response_queue_id)
         
-    
     for i in range( nprocesses ):
         
         # Prepate a new message
@@ -51,12 +49,15 @@ def parallel_execution( nprocesses, someparameter ):
         
     result = []
     for i in range( nprocesses ):
-        rs = q.get_messages()
-        m = rs[0]
-        resp = json.loads( m.get_body() )
-        q.delete_message(m)
+        # We expect the process to complete in 3 minutes or less
+        rs = q.get_messages(num_messages=1, wait_time_seconds=20)
         
-        result.append(resp['output'])
+        if ( len(rs) > 0 ):
+            m = rs[0]
+            resp = json.loads( m.get_body() )
+            q.delete_message(m)
+        
+            result.append(resp['output'])
     
     # We are done, delete response q
     q.delete()
